# apps/api/openapi/openapi.json
{
...
  "info": {
    "title": "Cloud Forest API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/v1/session": {
      "get": {
        "operationId": "getCurrentSessionV1",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "apiVersion": {
                      "enum": [
                        "v1"
                      ],
                      "type": "string"
                    },
                    "data": {
                      "additionalProperties": false,
                      "properties": {
                        "currentPersonId": {
                          "maxLength": 128,
                          "minLength": 1,
                          "type": "string"
                        },
                        "role": {
                          "anyOf": [
                            {
                              "enum": [
                                "admin"
                              ],
                              "type": "string"
                            },
                            {
                              "enum": [
                                "user"
                              ],
                              "type": "string"
                            }
                          ]
                        }
                      },
                      "required": [
                        "currentPersonId"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "apiVersion",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "apiVersion": {
                      "enum": [
                        "v1"
                      ],
                      "type": "string"
                    },
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "UNAUTHORIZED"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "enum": [
                            "A valid invited session is required."
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "apiVersion",
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "Session"
        ]
      }
    },
    "/api/v1/session/logout": {
      "post": {
        "operationId": "logoutV1",
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "apiVersion": {
                      "enum": [
                        "v1"
                      ],
                      "type": "string"
                    },
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "UNAUTHORIZED"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "enum": [
                            "A valid invited session is required."
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "apiVersion",
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "Session"
        ]
      }
    },
    "/api/v1/signup": {
      "post": {
        "operationId": "signupV1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "code": {
                    "maxLength": 256,
                    "minLength": 1,
                    "type": "string"
                  },
                  "password": {
                    "maxLength": 200,
                    "minLength": 8,
                    "type": "string"
                  },
                  "username": {
                    "maxLength": 30,
                    "minLength": 3,
                    "type": "string"
                  }
                },
                "required": [
                  "code",
                  "username",
                  "password"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Default Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "properties": {
                        "apiVersion": {
                          "enum": [
                            "v1"
                          ],
                          "type": "string"
                        },
                        "error": {
                          "properties": {
                            "code": {
                              "enum": [
                                "SIGNUP_FAILED"
                              ],
                              "type": "string"
                            },
                            "message": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "code",
                            "message"
                          ],
                          "type": "object"
                        }
                      },
                      "required": [
                        "apiVersion",
                        "error"
                      ],
                      "type": "object"
                    },
                    {
                      "type": "string"
                    }
                  ]
                }
              }
            },
            "description": "Default Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "apiVersion": {
                      "enum": [
                        "v1"
                      ],
                      "type": "string"
                    },
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "SIGNUP_FAILED"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "apiVersion",
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "Signup"
        ]
      }
    },
    "/api/v1/signup-codes": {
      "post": {
        "operationId": "createSignupCodeV1",
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "apiVersion": {
                      "enum": [
                        "v1"
                      ],
                      "type": "string"
                    },
                    "data": {
                      "properties": {
                        "link": {
                          "minLength": 1,
                          "type": "string"
                        }
                      },
                      "required": [
                        "link"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "apiVersion",
                    "data"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "apiVersion": {
                      "enum": [
                        "v1"
                      ],
                      "type": "string"
                    },
                    "error": {
                      "additionalProperties": false,
                      "properties": {
                        "code": {
                          "enum": [
                            "UNAUTHORIZED"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "enum": [
                            "A valid invited session is required."
                          ],
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "apiVersion",
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "apiVersion": {
                      "enum": [
                        "v1"
                      ],
                      "type": "string"
                    },
                    "error": {
                      "properties": {
                        "code": {
                          "enum": [
                            "FORBIDDEN"
                          ],
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "apiVersion",
                    "error"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        },
        "tags": [
          "Signup"
        ]
      }
    },
...
  }
}